Bug fix: Method ID Value generation was assuming all methods of all base classes had already been processed.
The invalid assumption described above appears to have always held true under MSBuild in VS2013 because of the order of the methods listed in the assembly (or at least, the order returned by Reflection). Under the latest MSBuild for VS2015, the order of methods changed thus violating the assumption for the first time and revealing the bug. The bug came as a callvirt instruction was resulting in a call to the wrong method (due to conflicting/duplicate method ID Values) and thus the called method was generating GC errors as what it thought was the "this" object reference was actually a different argument for a different method. Notably, had the mixed-up method and the intended method had the same number of arguments (inc. return value as an argument if not of type void), then bug would not have been detected until strange behaviour was observed, as opposed to the actual error raised by the GC.