Started transforming File System code.
Started transforming the file system code so that it uses storage pipes to connect to storage controllers (which are essentially controllers for managing access to one or more disks by one or more other processes). 

WaitOnPipeCreate now uses a request structure and returns the Id of the remote (i.e. input) process for the newly created pipe.

Added StorageData and StorageCommand Inpoint/Outpoint classes.

Turned off EHCI and HCI trace.