Resolves issue #86
Fixed stack state tracking in the compiler. The stack is now forked whenever a branch instruction is encountered/processed. Note: Changes have not yet been ported to MIPS.

Also, switched off the Deferred File Systems stuff so that the kernel is a bit more stable. Expect it to be able to run until it attempts to do an InitFS call. Then all hell breaks loose because of the (expected) null reference exception.