Special mapping for kernel process page
Added special tracking and mapping functions to virtual memory handling so that virtual pages for the Kernel Process are never used by any other address space (otherwise conflicts occur). Will need to add extra code in future to track when a page is no longer used by any process in order for it to become available to the kernel process.