Update to StatFS
FlingOS considers everything as pipes (unlike Linux which views everything as files). So StatFS now returns the process Id of the process which owns each FS mapping so that a process can connect to it and access the file system.

Also fixed a memory leak in the System Status Task.