Big system performance boost. Abort blocking R/W on pipes system call added
System call for aborting blocking read/write calls on a pipe now implemented.
This allows a separate thread of the original caller process of the blocking call to abort the call.
This allows Window Manager to use blocking calls (which massively reduces system overheads because it was using non-blocking calls in an endless loop which was obviously bad). It also makes Alt+Tab a bit more responsive and reliable because Alt+Tab is synchronised with the main ('display') thread properly.