Keyboard performance optimisation
Pre-load the scancodes for the "interesting" keys then just search those small sets for detecting if the key shiould be handled by Window Manager or passed to the current window. End result is much less time spent searching in WM thread and so higher throughput to the application. Overall this brings a massive performance increase to typing in FlingOS.