Cleaning up and demonstrating access of file systems and directories
All access (from an application's point of view) should be done through the FileSystemAccessor class. The static StatFSBySysCalls allows the application to get a list of file system mappings and the respective process ids of the file system manager processes. This allows it to create one file system accessor per target file system manager (process). One process can be managing multiple file systems and thus multiple mappings. 

An instance of a FileSystemAccessor then allows access to al the mappings that the connected FileSystemManager is managing. This in turn allows listing of directories (and eventually, file create/read/write/resize/delete).